requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
openai>=1.10.0
python-dotenv>=1.0.0
jinja2>=3.1.0
weasyprint>=60.0
chardet>=5.0.0
xlrd>=2.0.0
